refactor(engine): extract createCoordinatorDeps and fix circular dep
